refactor: 위치 입력 형식 변경
