+++ src/main/java/view/InputParser.java
@@ -1,13 +1,14 @@
-package domain;
+package view;
 
-
+import domain.Position;
 import domain.player.Name;
 import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Pattern;
 
 public class InputParser {
-    public static final Pattern POSITION_PATTERN = Pattern.compile(" *\\( *\\d+ *, *\\d+ *\\) *");
+    private static final Pattern COMMA_SEPARATED_COORDINATES = Pattern.compile("^\\s*\\d+\\s*,\\s*\\d+\\s*$");
+    private static final String DELIMITER = ",";
 
     public static Name parseName(String input) {
         if (input == null || input.isBlank()) {
@@ -20,16 +21,16 @@ public class InputParser {
         if (input == null || input.isBlank()) {
             throw new IllegalArgumentException("빈 값은 입력할 수 없습니다.");
         }
-        
-        if (!POSITION_PATTERN.matcher(input).matches()) {
-            throw new IllegalArgumentException("질못된 입력 형식입니다.");
+
+        if (!COMMA_SEPARATED_COORDINATES.matcher(input).matches()) {
+            throw new IllegalArgumentException("잘못된 입력 형식입니다. (예: 0,3)");
         }
 
         return getPosition(input);
-   }
+    }
 
     private static Position getPosition(String input) {
-        List<Integer> coordinate = Arrays.stream(input.strip().substring(1, input.length() - 1).split(","))
+        List<Integer> coordinate = Arrays.stream(input.split(DELIMITER))
                 .map(String::strip)
                 .map(Integer::parseInt)
                 .toList();
