+++ src/test/java/domain/board/FormationCommandTest.java
@@ -10,7 +10,7 @@ class FormationCommandTest {
     @Test
     void 포메이션_입력이_1_2_3_4_일때_정상_동작한다() {
         assertThat(FormationCommand.from("1")).isEqualTo(FormationCommand.FIRST);
-        assertThat(FormationCommand.from(" 2 ")).isEqualTo(FormationCommand.SECOND);
+        assertThat(FormationCommand.from("2")).isEqualTo(FormationCommand.SECOND);
         assertThat(FormationCommand.from("3")).isEqualTo(FormationCommand.THIRD);
         assertThat(FormationCommand.from("4")).isEqualTo(FormationCommand.FOURTH);
     }
